Make napi_env live as long as NapiEnv instead of per-call stack locals
Move napi_env__ into a NapiEnv member (NapiEnv::env()) so a napi_callback
captured by something long-lived (e.g. a napi_wrap finalizer) never captures
a dangling env once the Evaluator::execute call that created it returns.
Simplifies all cctest napi_env__ boilerplace down to napiEnv->env().

This is a distinct fix from the "flush wrapped objects via forced GC before
a test ends" pattern already used in ObjectWrap/FactoryWrap - both are still
needed, since leftover GC garbage can otherwise get finalized at an unsafe
moment (verified via gdb: mid-construction of an unrelated test's VMInstance).
